# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: R&S Admissao API
  description: |
    API do sistema de Recrutamento e Selecao (R&S) para agendamento de aulas teste.

    Este sistema permite que candidatos agendem aulas teste em unidades educacionais,
    enquanto a equipe de recrutamento gerencia vagas, configuracoes de horarios e
    bloqueios de agenda.

    ## Autenticacao

    A API utiliza dois mecanismos de autenticacao:

    - **JWT Bearer Token**: Para usuarios autenticados (equipe de recrutamento)
    - **Acesso Publico**: Alguns endpoints sao publicos para candidatos

    Endpoints marcados com `optionalAuth` podem ser acessados com ou sem autenticacao,
    com comportamento diferenciado baseado no contexto.

    ## Ambientes

    | Ambiente | URL Base |
    |----------|----------|
    | Producao | https://api.rsadmissao.tomeducacao.com.br/api/v1 |
    | Homologacao | https://api-homolog.rsadmissao.tomeducacao.com.br/api/v1 |

    ## Versionamento

    A API segue versionamento semantico. A versao atual e `v1`.

    ## Rate Limiting

    A API implementa limitacao de taxa para prevenir abuso:

    | Endpoint | Limite | Janela |
    |----------|--------|--------|
    | `POST /auth/login` | 5 requisicoes | 1 minuto |
    | `POST /booking` (publico) | 10 requisicoes | 1 minuto |
    | `GET /availability` | 60 requisicoes | 1 minuto |
    | Demais endpoints | 100 requisicoes | 1 minuto |

    **Headers de resposta:**
    - `X-RateLimit-Limit`: Limite total de requisicoes
...

...

...

...

...
servers:
...
paths:
...
  /availability:
    get:
      tags:
        - availability
      summary: Consultar slots disponiveis
      description: |
        Retorna os slots disponiveis para agendamento de aulas teste.

        O comportamento varia conforme o contexto de autenticacao:

        | Aspecto | Publico (sem auth) | Privado (com auth) |
        |---------|--------------------|--------------------|
        | Range de datas | Menor entre d_rules e valid_from/valid_until | valid_from ate valid_until |
        | Sabado | Excluido | Incluido |
        | Status dos slots | Apenas `vago` | Todos (vago, ocupado, bloqueado) |
        | Config exposta | Apenas slot_size | Configuracao completa |

        A paginacao e feita por semana, usando domingo como separador.
      operationId: getAvailability
      security:
        - {}
        - bearerAuth: []
      parameters:
        - name: id_unidade
          in: query
          required: true
          description: ID da unidade para consultar disponibilidade
          schema:
            type: integer
            example: 123
        - name: page
          in: query
          required: false
          description: |
            Numero da pagina (semana) a ser retornada.
            Se omitido, retorna a semana que contem a data atual.
          schema:
            type: integer
            minimum: 1
            example: 1
      responses:
        '200':
          description: Slots disponiveis retornados com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilityResponse'
              examples:
                publico:
                  summary: Resposta para acesso publico
                  value:
                    success: true
                    config:
                      slot_size: 50
                    pagination:
                      current_page: 1
                      total_pages: 4
                      week_start: "2025-01-06"
                      week_end: "2025-01-10"
                    slots:
                      - date: "2025-01-06"
                        day_of_week: "segunda-feira"
                        slot_start: "08:00:00"
                        slot_end: "08:50:00"
                        status: "vago"
                      - date: "2025-01-06"
                        day_of_week: "segunda-feira"
                        slot_start: "09:00:00"
                        slot_end: "09:50:00"
                        status: "vago"
                privado:
                  summary: Resposta para acesso autenticado
                  value:
                    success: true
                    config:
                      slot_size: 50
                      morning_start_at: "08:00:00"
                      morning_end_at: "12:00:00"
                      afternoon_start_at: "13:00:00"
                      afternoon_end_at: "17:00:00"
                      d_rule_start: 1
                      d_rule_end: 30
                      valid_from: "2025-01-01"
                      valid_until: "2025-03-31"
                    pagination:
                      current_page: 1
                      total_pages: 4
                      week_start: "2025-01-06"
                      week_end: "2025-01-11"
                    slots:
                      - date: "2025-01-06"
                        day_of_week: "segunda-feira"
                        slot_start: "08:00:00"
                        slot_end: "08:50:00"
                        status: "vago"
                      - date: "2025-01-06"
                        day_of_week: "segunda-feira"
                        slot_start: "09:00:00"
                        slot_end: "09:50:00"
                        status: "ocupado"
                        booking_id: 456
                        candidate_name: "Maria Silva"
        '400':
          description: Parametros invalidos ou periodo nao disponivel
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missingIdUnidade:
                  summary: id_unidade nao informado
                  value:
                    success: false
                    errors:
                      - msg: "id_unidade e obrigatorio"
                        param: "id_unidade"
                        location: "query"
                periodoIndisponivel:
                  summary: Periodo de agendamento nao disponivel
                  value:
                    success: false
                    error: "Periodo de agendamento nao disponivel"
        '404':
          description: Configuracao de horarios nao encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Configuracao de horarios nao encontrada"
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ============================================================================
  # BOOKING
  # ============================================================================
...
  /schedule:
    get:
      tags:
        - schedule
      summary: Consultar configuracao de horarios
      description: |
        Retorna a configuracao de horarios para uma unidade ou a configuracao global.

        **Hierarquia de configuracao:**
        1. Se `id_unidade` for informado e existir config especifica, retorna ela
        2. Se `id_unidade` for informado mas nao houver config especifica, retorna global
        3. Se `id_unidade` nao for informado, retorna global

        Requer autenticacao com role `recrutamento` ou `admin`.
      operationId: getScheduleConfig
      security:
        - bearerAuth: []
      parameters:
        - name: id_unidade
          in: query
          required: false
          description: |
            ID da unidade para buscar configuracao especifica.
            Se omitido, retorna a configuracao global.
          schema:
            type: integer
            example: 123
      responses:
        '200':
          description: Configuracao retornada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleConfigResponse'
              examples:
                global:
                  summary: Configuracao global
                  value:
                    success: true
                    config:
                      id_schedule_config: 1
                      id_unidade: null
                      morning_start_at: "08:00:00"
                      morning_end_at: "12:00:00"
                      afternoon_start_at: "13:00:00"
                      afternoon_end_at: "17:00:00"
                      slot_size: 50
                      d_rule_start: 1
                      d_rule_end: 30
                      valid_from: null
                      valid_until: null
                      active: true
                    is_global: true
                unidade:
                  summary: Configuracao especifica de unidade
                  value:
                    success: true
                    config:
                      id_schedule_config: 5
                      id_unidade: 123
                      morning_start_at: "07:30:00"
                      morning_end_at: "11:30:00"
                      afternoon_start_at: "13:30:00"
                      afternoon_end_at: "17:30:00"
                      slot_size: 45
                      d_rule_start: 2
                      d_rule_end: 15
                      valid_from: "2025-01-01"
                      valid_until: "2025-06-30"
                      active: true
                    is_global: false
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Configuracao nao encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Configuracao de horarios nao encontrada"
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - schedule
      summary: Criar ou atualizar configuracao de horarios
      description: |
        Cria ou atualiza a configuracao de horarios.

        **Comportamento:**
        - Se `id_unidade` for `null` ou omitido, afeta a configuracao global
        - Se `id_unidade` for informado, cria/atualiza config especifica da unidade
        - Operacao de upsert: cria se nao existir, atualiza se existir

        Requer autenticacao com role `recrutamento` ou `admin`.
      operationId: upsertScheduleConfig
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleConfigRequest'
            examples:
              global:
                summary: Configuracao global
                value:
                  id_unidade: null
                  morning_start_at: "08:00:00"
                  morning_end_at: "12:00:00"
                  afternoon_start_at: "13:00:00"
                  afternoon_end_at: "17:00:00"
                  slot_size: 50
                  d_rule_start: 1
                  d_rule_end: 30
              unidade:
                summary: Configuracao especifica de unidade
                value:
                  id_unidade: 123
                  morning_start_at: "07:30:00"
                  morning_end_at: "11:30:00"
                  afternoon_start_at: "13:30:00"
                  afternoon_end_at: "17:30:00"
                  slot_size: 45
                  d_rule_start: 2
                  d_rule_end: 15
                  valid_from: "2025-01-01"
                  valid_until: "2025-06-30"
      responses:
        '200':
          description: Configuracao atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleConfigResponse'
              example:
                success: true
                message: "Configuracao atualizada com sucesso"
                config:
                  id_schedule_config: 5
                  id_unidade: 123
                  morning_start_at: "07:30:00"
                  morning_end_at: "11:30:00"
                  afternoon_start_at: "13:30:00"
                  afternoon_end_at: "17:30:00"
                  slot_size: 45
                  d_rule_start: 2
                  d_rule_end: 15
                  valid_from: "2025-01-01"
                  valid_until: "2025-06-30"
                  active: true
        '201':
          description: Configuracao criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleConfigResponse'
              example:
                success: true
                message: "Configuracao criada com sucesso"
                config:
                  id_schedule_config: 6
                  id_unidade: 456
                  morning_start_at: "08:00:00"
                  morning_end_at: "12:00:00"
                  afternoon_start_at: "13:00:00"
                  afternoon_end_at: "17:00:00"
                  slot_size: 50
                  d_rule_start: 1
                  d_rule_end: 30
                  valid_from: null
                  valid_until: null
                  active: true
        '400':
          description: Dados invalidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                errors:
                  - msg: "slot_size deve estar entre 15 e 120 minutos"
                    param: "slot_size"
                    location: "body"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtido atraves do endpoint `/auth/login`.

        Enviar no header: `Authorization: Bearer {token}`
  responses:
    Unauthorized:
      description: Token nao fornecido ou invalido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Token nao fornecido ou invalido"
    Forbidden:
      description: Usuario nao tem permissao para esta acao
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Acesso negado. Permissao insuficiente."
    InternalServerError:
      description: Erro interno do servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Erro interno do servidor"
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Mensagem de erro
          example: "Descricao do erro"
        errors:
          type: array
          description: Lista de erros de validacao
          items:
            type: object
            properties:
              msg:
                type: string
                example: "Campo obrigatorio"
              param:
                type: string
                example: "id_unidade"
              location:
                type: string
                example: "query"
        details:
          type: string
          description: Detalhes adicionais do erro (desenvolvimento)

    # --------------------------------------------------------------------------
    # Availability Schemas
    # --------------------------------------------------------------------------
    AvailabilityResponse:
      type: object
      required:
        - success
        - config
        - pagination
        - slots
      properties:
        success:
          type: boolean
          example: true
        config:
          $ref: '#/components/schemas/AvailabilityConfig'
        pagination:
          $ref: '#/components/schemas/WeekPagination'
        slots:
          type: array
          items:
            $ref: '#/components/schemas/Slot'
    AvailabilityConfig:
      type: object
      required:
        - slot_size
      properties:
        slot_size:
          type: integer
          description: Duracao do slot em minutos
          example: 50
        morning_start_at:
          type: string
          format: time
          description: Horario de inicio do periodo da manha (apenas para autenticados)
          example: "08:00:00"
        morning_end_at:
          type: string
          format: time
          description: Horario de fim do periodo da manha (apenas para autenticados)
          example: "12:00:00"
        afternoon_start_at:
          type: string
          format: time
          description: Horario de inicio do periodo da tarde (apenas para autenticados)
          example: "13:00:00"
        afternoon_end_at:
          type: string
          format: time
          description: Horario de fim do periodo da tarde (apenas para autenticados)
          example: "17:00:00"
        d_rule_start:
          type: integer
          description: Dias minimos de antecedencia para agendamento (apenas para autenticados)
          example: 1
        d_rule_end:
          type: integer
          description: Dias maximos de antecedencia para agendamento (apenas para autenticados)
          example: 30
        valid_from:
          type: string
          format: date
          nullable: true
          description: Data de inicio da validade da configuracao (apenas para autenticados)
          example: "2025-01-01"
        valid_until:
          type: string
          format: date
          nullable: true
          description: Data de fim da validade da configuracao (apenas para autenticados)
          example: "2025-03-31"
    WeekPagination:
      type: object
      required:
        - current_page
        - total_pages
        - week_start
        - week_end
      properties:
        current_page:
          type: integer
          description: Pagina (semana) atual
          minimum: 1
          example: 1
        total_pages:
          type: integer
          description: Total de paginas (semanas) disponiveis
          minimum: 0
          example: 4
        week_start:
          type: string
          format: date
          description: Data de inicio da semana atual
          example: "2025-01-06"
        week_end:
          type: string
          format: date
          description: Data de fim da semana atual
          example: "2025-01-10"
    Slot:
      type: object
      required:
        - date
        - day_of_week
        - slot_start
        - slot_end
        - status
      properties:
        date:
          type: string
          format: date
          description: Data do slot
          example: "2025-01-06"
        day_of_week:
          type: string
          description: Dia da semana em portugues
          example: "segunda-feira"
        slot_start:
          type: string
          format: time
          description: Horario de inicio do slot
          example: "08:00:00"
        slot_end:
          type: string
          format: time
          description: Horario de fim do slot
          example: "08:50:00"
        status:
          $ref: '#/components/schemas/SlotStatus'
        booking_id:
          type: integer
          description: ID do agendamento (apenas para slots ocupados, apenas para autenticados)
          example: 456
        candidate_name:
          type: string
          maxLength: 200
          description: Nome do candidato (apenas para slots ocupados, apenas para autenticados)
          example: "Maria Silva"
    SlotStatus:
      type: string
      enum:
        - vago
        - ocupado
        - bloqueado
      description: |
        Status do slot:
        - `vago`: Disponivel para agendamento
        - `ocupado`: Ja existe agendamento neste horario
        - `bloqueado`: Bloqueado por configuracao (schedule_block)

    # --------------------------------------------------------------------------
    # Booking Schemas
    # --------------------------------------------------------------------------
...
    ScheduleConfig:
      type: object
      required:
        - id_schedule_config
        - morning_start_at
        - morning_end_at
        - afternoon_start_at
        - afternoon_end_at
        - slot_size
        - d_rule_start
        - d_rule_end
        - active
      properties:
        id_schedule_config:
          type: integer
          description: ID da configuracao
          example: 1
        id_unidade:
          type: integer
          nullable: true
          description: ID da unidade (null = configuracao global)
          example: 123
        morning_start_at:
          type: string
          format: time
          description: Horario de inicio do periodo da manha
          example: "08:00:00"
        morning_end_at:
          type: string
          format: time
          description: Horario de fim do periodo da manha
          example: "12:00:00"
        afternoon_start_at:
          type: string
          format: time
          description: Horario de inicio do periodo da tarde
          example: "13:00:00"
        afternoon_end_at:
          type: string
          format: time
          description: Horario de fim do periodo da tarde
          example: "17:00:00"
        slot_size:
          type: integer
          minimum: 15
          maximum: 120
          description: Duracao do slot em minutos
          example: 50
        d_rule_start:
          type: integer
          minimum: 0
          description: Dias minimos de antecedencia para agendamento
          example: 1
        d_rule_end:
          type: integer
          minimum: 1
          description: Dias maximos de antecedencia para agendamento
          example: 30
        valid_from:
          type: string
          format: date
          nullable: true
          description: Data de inicio da validade da configuracao
          example: "2025-01-01"
        valid_until:
          type: string
          format: date
          nullable: true
          description: Data de fim da validade da configuracao
          example: "2025-03-31"
        active:
          type: boolean
          description: Indica se a configuracao esta ativa
          example: true
    ScheduleConfigRequest:
      type: object
      required:
        - morning_start_at
        - morning_end_at
        - afternoon_start_at
        - afternoon_end_at
        - slot_size
        - d_rule_start
        - d_rule_end
      properties:
        id_unidade:
          type: integer
          nullable: true
          description: ID da unidade (null = configuracao global)
          example: 123
        morning_start_at:
          type: string
          format: time
          description: Horario de inicio do periodo da manha
          example: "08:00:00"
        morning_end_at:
          type: string
          format: time
          description: Horario de fim do periodo da manha
          example: "12:00:00"
        afternoon_start_at:
          type: string
          format: time
          description: Horario de inicio do periodo da tarde
          example: "13:00:00"
        afternoon_end_at:
          type: string
          format: time
          description: Horario de fim do periodo da tarde
          example: "17:00:00"
        slot_size:
          type: integer
          minimum: 15
          maximum: 120
          description: Duracao do slot em minutos
          example: 50
        d_rule_start:
          type: integer
          minimum: 0
          description: Dias minimos de antecedencia para agendamento
          example: 1
        d_rule_end:
          type: integer
          minimum: 1
          description: Dias maximos de antecedencia para agendamento
          example: 30
        valid_from:
          type: string
          format: date
          nullable: true
          description: Data de inicio da validade da configuracao
          example: "2025-01-01"
        valid_until:
          type: string
          format: date
          nullable: true
          description: Data de fim da validade da configuracao
          example: "2025-03-31"
        active:
          type: boolean
          default: true
          description: Indica se a configuracao esta ativa
          example: true
    ScheduleConfigResponse:
      type: object
      required:
        - success
        - config
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Configuracao atualizada com sucesso"
        config:
          $ref: '#/components/schemas/ScheduleConfig'
        is_global:
          type: boolean
          description: Indica se a configuracao retornada e global
          example: false

    # --------------------------------------------------------------------------
    # Schedule Block Schemas
    # --------------------------------------------------------------------------
...